Update KMIP client certificate authentication process
Based on the documentation, forcing TLS 1.3 with Pythons `ssl` module disables
automatic client certificate authentication, instead relying on the
`verify_client_post_handshake` method of the server. However, this
introduces an extra round trip. However, not explicitly setting TLS
version and requiring client authentication will apparently cause OpenSSL to
use TLS 1.3 (when available) and authenticating the client as part of
the handshake. This is somewhat confusing, and potentially problematic
as we don't have a guarantee that this TLS version will be used.
However, for these experiments it seems like a better choice, and in a
real application this could most likely be controlled properly.