Ensure gRPC server only uses a single thread
Since both the KMIP and REST APIs are stingle threaded, the gRPC API has
been changed to be single threaded as well. In general, the signing
interface is not thread safe, as we would need to make sure that no
concurrent interactions for the same session happens. While probably not
impossible to fix, it is easier to just restrict it to single threaded
access for now.