Allow REST server address reuse
Previouly we could run into the error "Address already in use" when we
restarted the REST API server since the socket used by the previous
server is still waiting for all packets to expire (the TIME-WAIT
state). By setting the `allow_reuse_address` property on the server
class this should now be fixed.