Update sleep function for MockHSM
The `MockHSM` previously used the `time.sleep` function internally to
wait for a constant amount of time when mocking a signature operation.
However, this method was tested and found to be quite inaccurate. In
particular, it was unable to sleep for very short amounts of time (such
as requited for a HSM capable of 20k mock signatures per second).
In order to remedy this, the `sign` function now uses a custom sleep
function based on `time.perf_counter`, which seems to allow for much
finer granularity.