Add locking for MockHSM when accessed from multiple threads
Since a real HSM has a hard limit on the number of signatures it can
perform per second, we must make sure we cannot circumvent this
in the mock HSM by performing signatures in concurrently. For this reason,
the `MockHSM` now has a `thread_safe` constructor parameter. When set,
the HSM will use a `threading.Lock` to make sure that the `sign`
operation can only be accessed from a tread at a time. This change was
done in preparation for adding multithreading to the API
implementations.